HBA Bedouine's Farm Small Project Finalized
This project utilizes Node.js to create a web server and perform various file system operations. It employs built-in modules like HTTP, URL, and File System, along with a third-party library, Slugify. The synchronous and asynchronous approaches are showcased for reading and writing files. Additionally, it demonstrates routing for different URL paths and serves HTML templates and API data. Finally, it listens for incoming requests on port 15100, enabling access to server resources.